requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.41.0
cryptography==44.0.0
Django==4.2.17
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
frozenlist==1.5.0
idna==3.10
generics==7.0.0
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
multidict==6.1.0
jwcrypto==1.5.6
mysqlclient==2.2.6
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
PyJWT==2.10.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
twilio==9.4.1
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
